Add redux
