+++ src/reducers/index.ts
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+
+const reducers = combineReducers({});
+
+export type RootState = ReturnType<typeof reducers>;
+export default reducers;
