+++ src/styles/Canvas.scss
@@ -1,4 +1,4 @@
-.App {
+.card-canvas {
   display: flex;
   justify-content: space-between;
   align-items: center;
